extract common logic to ForecastFragment, add background fragment
