+++ app/src/main/java/com/example/weather/fragments/WeatherDisplayFragment.kt
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
+import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
@@ -18,43 +19,21 @@ import com.example.weather.support.Handler
 import com.example.weather.viewmodels.ForecastViewModel
 
 
-class WeatherDisplayFragment() : Fragment() {
-    private val viewModel by lazy {
-        ViewModelProvider(this).get(ForecastViewModel::class.java)
-    }
-
-    private fun setupObservers() {
-        val forecastObserver = Observer<Forecast> { forecast -> updateScreen(forecast) }
-        viewModel.subscribeToForecast(this, forecastObserver)
-
-        val errorObserver = Observer<AppError> { error -> handleError(error) }
-        viewModel.subscribeToError(this, errorObserver)
-    }
-
+class WeatherDisplayFragment : ForecastFragment() {
     private fun setupView(view: View) {
         view.findViewById<Button>(R.id.bRefresh).setOnClickListener { fetchForecast() }
     }
 
-    private fun updateScreen(forecast: Forecast) {
+    override fun updateScreen(forecast: Forecast) {
         view?.findViewById<TextView>(R.id.weatherText)?.text = getString(R.string.temperature_as_celsius, forecast.temp.toInt())
         view?.findViewById<TextView>(R.id.rainText)?.text = getString(R.string.rain_probability, forecast.rain)
     }
 
-    private fun handleError(error: AppError) {
-        Handler.handleError(activity!!, error)
+    override fun handleError(error: AppError) {
+        super.handleError(error)
 
         view?.findViewById<TextView>(R.id.weatherText)?.text = DEFAULT_ERROR_VALUE
         view?.findViewById<TextView>(R.id.rainText)?.text = DEFAULT_ERROR_VALUE
-    }
-
-    private fun fetchForecast() {
-        AsyncTask.run { viewModel.fetchForecast() }
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setupObservers()
-        fetchForecast()
     }
 
     override fun onCreateView(
